diff_checker/requirements.txt
pytesseract
python-docx
pymupdf
requests
pdfplumber
python-docx
google-api-python-client  # If Gemini uses Google API Python Client
requests
